Supersedes #28 and fixes bugs found in Dagger PR https://github.com/JuliaParallel/Dagger.jl/pull/141