@jpsamaroo 2020-09-23
Also bumped version to 0.3.2 so that Dagger PRs can get `poolref`/`poolunref`