@jpsamaroo 2020-09-23
@shashi I'd appreciate if you can confirm that this is sufficiently thread-safe; should I call `with_datastore_lock()` to ensure safety?