@jpsamaroo 2020-09-23
I've made all new remote calls asynchronous. I'm not sure the async calls are 100% correct, since we should technically wait on them to ensure integrity, but I'd imagine that in the common case it won't be an issue.