@jpsamaroo 2020-10-05
The latest commit puts the `DRef` in the refcounting tests on worker 2 instead of 1 to illustrate that this actually works with remote references.