@jpsamaroo 2020-10-06
Benchmarks:

```julia
f() = [poolset(1) for i in 1:1000]
```

Master:
```julia
julia> @btime f()
  485.746 μs (5004 allocations: 179.83 KiB)
julia> @btime sort(da) setup=(da=compute(rand(Blocks(64),256)))
  2.824 ms (11210 allocations: 690.94 KiB)
julia> @btime sort(da) setup=(da=compute(rand(Blocks(16),2048)))
  4.908 s (33671300 allocations: 2.82 GiB)
julia> @btime sort(compute(rand(Blocks(16),2048)))
  4.812 s (33787355 allocations: 2.83 GiB)
```

This PR:
```julia
julia> @btime f()
  696.458 μs (7003 allocations: 195.50 KiB)
julia> @btime sort(da) setup=(da=compute(rand(Blocks(64),256)))
  2.831 ms (11339 allocations: 695.88 KiB)
julia> @btime sort(da) setup=(da=compute(rand(Blocks(16),2048)))
  4.582 s (33761248 allocations: 2.82 GiB)
julia> @btime sort(compute(rand(Blocks(16),2048)))
  4.903 s (33859632 allocations: 2.83 GiB)
```

@shashi benchmarks look pretty decent!