@jpsamaroo 2020-10-19
Ok, here's an edge case this doesn't handle: worker A holds the only reference to object X. Worker A sends the reference to worker B, and then immediately GC's the reference, sending an unref message to the owner. While worker B is working on deserializing the reference, the owner receives the unref message and immediately `pooldelete`s the object, causing a `KeyError` on worker B when it eventually attempts to access the object. Also, the `poolref` attempted on worker B will quietly error, since it's never waited on.